+++ index.php
@@ -20,9 +20,9 @@
  * @subpackage         	: Main Frontend file
  * @source             	: /[BLOCKS PATH]/phpclasses_latest/index.php
  * @fileNo             	: 
- * @version            	: 1.0.0
+ * @version            	: 1.0.1
  * @created            	: 2024-10-10 07:00:00 UTC+3
- * @updated            	: 
+ * @updated            	: 2024-10-14 07:00:00 UTC+3
  * @author             	: Drogidis Christos
  * @authorSite         	: www.alexsoft.gr
  * @license 			: AGL-F
@@ -32,17 +32,58 @@
 
 defined ("ALEXSOFT_RUN_CMS") or die("Prohibition of Access.");
 
+include_once ( __DIR__ . '/libs.php');
+
 
 // Get Value Block Parameters
-$count 		= $block->getParam('int', 'count', 10 );					// How many classes will be displayed.
-$user		= $block->getParam('str', 'username', 'bigfriend' );	// PHPClasses User
-//$show_days	= $block->getParam('bool', 'show_days', false );		// Show Days
-$theme		= $block->getParam('lstr', 'theme', 'default' );		// The Block theme
+$method 		= $block->getParam('str', 'method', 'xml' );			// What feed method is used (XML or RSS)?
+$count 			= $block->getParam('int', 'count', 10 );				// How many classes will be displayed.
+$user			= $block->getParam('str', 'username', 'bigfriend' );	// PHPClasses User
+$show_summary	= $block->getParam('bool', 'show_summary', false );		// Show Summary Description
+$show_days		= $block->getParam('bool', 'show_days', false );		// Show Days
+$theme			= $block->getParam('lstr', 'theme', 'default' );		// The Block theme
 
 // load Block Theme
 $block->loadTheme($theme);
 
-$xml = simplexml_load_file('https://'.$user.'.users.phpclasses.org/browse/latest/latest.xml');
+$feeds = simplexml_load_file('https://'.$user.'.users.phpclasses.org/browse/latest/latest.'.$method);
+
+$items= [];
+
+switch ($method) {
+	case 'rss':
+		$ci = 0;
+
+		foreach ($feeds->channel->item as $feed) {
+			if ($ci >= $count) break;	
+			$parts = explode("#", $feed->link);
+			$items[$ci]['link'] = $parts[0];
+			$items[$ci]['title'] = (string) $feed->title;
+			if ($show_summary) $items[$ci]['summary'] = getSummary($feed->description);	
+			if ($show_days) $items[$ci]['days'] = diff_days($feed->pubDate);		
+			
+			$ci++;
+			unset($parts);
+		}
+		unset($feeds);
+		break;
+		
+	default:
+		$ci = 0;
+
+		foreach ($feeds->item as $feed) {
+			if ($ci >= $count) break;	
+			$parts = explode("#", $feed->link);
+			$items[$ci]['link'] = $parts[0];
+			$items[$ci]['title'] = (string) $feed->title;
+			if ($show_summary) $items[$ci]['summary'] = getSummary($feed->description);	
+			
+			$ci++;
+			unset($parts);
+		}
+		unset($feeds);
+		break;
+}
 
 
 $text = '';
@@ -52,20 +93,20 @@ if ($block->getVar('show_title')) {
 }
 $text .= '<div class="text"><div class="table">';
 
-$ci =1;
 
-foreach ($xml->item as $feed) {
-	if ($ci > $count) break;
-
-	$parts = explode("#", $feed->link);
+foreach ($items as $key => $feed)
+{
 	$text .= '<div class="row">';
-	$text .= '<div class="cell"><a target="_blank" href="'.$parts[0].'">'.$feed->title.'</a></div>';
-	$text .= '</div>'; // row	
+		$text .= '<div class="cell"><a target="_blank" href="'.$feed['link'].'">'.$feed['title'].'</a>';
+			if ($show_summary) $text .= '<br>'.$feed['summary'];
+		$text .= '</div>';
 	
-	$ci++;
-	unset($parts);
+		// Only for RSS Feed
+		if ($show_days && $method !== 'xml') $text .= "<div class=\"cell right\">".$feed['days']."</div>";	
+	
+	$text .= '</div>'; // row
 }
-unset($xml);
+unset($items);
 
 $text .= '</div></div>'; // table/text
 $text .= '<div class="more"><a target="_blank" href="https://'.$user.'.users.phpclasses.org/browse/latest/latest.html"><strong>...'.$objLang->more.'</strong></a></div>';