+++ languages/el-CY.php
@@ -20,9 +20,9 @@
  * @subpackage         	: Greek Language file
  * @source             	: /[BLOCKS PATH]/phpclasses_latest/languages/el.php
  * @fileNo             	: 
- * @version            	: 1.0.0
+ * @version            	: 1.0.1
  * @created            	: 2024-10-10 07:00:00 UTC+3
- * @updated            	: 
+ * @updated            	: 2024-10-14 07:00:00 UTC+3
  * @author             	: Drogidis Christos
  * @authorSite         	: www.alexsoft.gr
  * @translator         	:
@@ -48,10 +48,17 @@ class TBlock_phpclasses_latest_Language extends TObject
 
 	public $APL_paramgroup_general_label = "▼ &nbsp; Επιλογές Γενικών Παραμέτρων";
 	public $APL_paramgroup_general_label_info = "<strong>ΕΠΙΛΟΓΕΣ ΓΕΝΙΚΩΝ ΠΑΡΑΜΕΤΡΩΝ</strong><br />--------------------------------------<br /><br />Στο πλαίσιο αυτό μπορείτε να επιλέξετε από διάφορες γενικές παραμέτρους και να διαμορφώσετε δυναμικά την ενότητα";
+	public $APL_method_label = "Επιλογή αρχείου τροφοδοσίας";
+	public $APL_method_placeholder = " Επιλέξτε Feed (XML ή RSS)";
+	public $APL_method_desc = "<strong>ΕΠΙΛΟΓΗ ΑΡΧΕΙΟΥ ΤΡΟΦΟΔΟΣΙΑΣ</strong><br />--------------------------------------<br /><br />Εδώ πρέπει να επιλέξετε την μορφή του αρχείου τροφοδοσίας. Έχετε δύο επιλογές, XML ή RSS.";	
+	public $APL_opt_XML = "XML";
+	public $APL_opt_RSS = "RSS 2.0";
 	public $APL_count_label = "Αριθμός κλάσεων";
 	public $APL_count_desc = "<strong>ΑΡΙΘΜΟΣ ΚΛΑΣΕΩΝ</strong><br />--------------------------------------<br /><br />Εδώ πρέπει να επιλέξετε τον αριθμό των κλάσεων που θα εμφανιστούν";
 	public $APL_username_label = "Όνομα Χρήστη";
 	public $APL_username_desc = "<strong>ΟΝΟΜΑ ΧΡΗΣΤΗ</strong><br />--------------------------------------<br /><br />Εδώ πρέπει να δώσετε το αναγνωριστικό όνομα χρήστη που συνδέεστε στο phpclasses.org";
+	public $APL_show_summary_label = "Εμφάνιση σύντομης περιγραφής";
+	public $APL_show_summary_desc = "<strong>ΕΜΦΑΝΙΣΗ ΣΥΝΤΟΜΗΣ ΠΕΡΙΓΡΑΦΗΣ</strong><br />--------------------------------------<br /><br />Θέλετε να εμφανίζεται μια σύντομη περιγραφή για κάθε σύνδεσμο;";
 	public $APL_show_days_label = "Εμφάνιση Ημερών";
 	public $APL_show_days_desc = "<strong>ΕΜΦΑΝΙΣΗ ΗΜΕΡΩΝ</strong><br />--------------------------------------<br /><br />Θέλετε να εμφανίζεται στήλη με τις ημέρες που πέρασαν από την ανάρτηση της κλάσης;";
 	public $APL_theme_label = "Θέμα εμφάνισης Ενότητας";