+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.1] - 2024-10-14
+
+- Added support for both two feed methods (XML or RSS)
+- Added support for description summary 
+- Added support for show days that have passed since the package was published. ONLY RSS Feed. 
+
 ## [1.0.0] - 2024-10-10
 
 - Initial release